tests/wpt/web-platform-tests/old-tests/submission/Microsoft/selection/collapseToEnd.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Collapse the selection with collapseToEnd()</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var selection = window.getSelection();
                var p1 = document.getElementById("p1");

                var range = document.createRange();
                range.selectNode(p1);
                selection.addRange(range);

                checkSelectionAttributes(document.body, 1, document.body, 2, false, 1);
                assert_equals(selection.toString(), p1.textContent);

                selection.collapseToEnd();

                checkSelectionAttributes(document.body, 2, document.body, 2, true, 1);
                assert_equals(selection.toString(), "");
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <p id="p1">Collapse the selection with collapseToEnd()</p>
    </body>
</html>
